Continuous Extreme Value Models

Contrary to the discrete extreme value models, continuous extreme value models require covariates and thus allow estimation of extreme events for evaluation sites with yet unobserved data.

SpatialBayesExtremes.ContinuousExtremeValueModel — Type

abstract type ContinuousExtremeValueModel <: ExtremeValueModel end

An abstract type for continuous extreme value models.

SpatialBayesExtremes.getDistribution — Method

getDistribution(model::ContinuousExtremeValueModel, covariates::DataFrame, θ::NamedTuple)

Returns the probability distribution associated with the model and parameters for an observation sites indicated by covariates.

Currently, the only available class of continuous extreme value models are regression extreme value models.

Regression Extreme Value Models

As in the univariate case, we assume that the response variable can be modeled using a univariate extreme value model, but its parameters are allowed to vary with covariates through regression representations:

Model setting

Conditional on the covariates, each observation is assumed to follow an extreme value distribution (e.g. GEV, GPD). All observations are assumed to be independent, irrespective of the covariates.

In contrast to stationary models with constant parameters, we allow each model parameter to depend on covariates. For a generic parameter $θ$, we assume $ \theta = f(x) $ where $x$ denotes a covariate vector (e.g. the covariates associated with some observation), and $f$ belongs to a user-specified function class as follows: Although th dependence on the covariates may be nonlinear, the model is linear in its unknown parameters. More precisely, each parameter function can is represented as $ \theta = \sum{i = 1}^{m} \beta{j} \, \phi_{j}(x) $ where:

\[\beta = (\beta_{1}, \dots, \beta_{m})\]
are regression coefficients

\[\phi_{j}\]
are functions of the covariates

The basis functions $\phi_{j}(\cdot)$ may include:

raw covariates

transformations such as powers or interactions

user-defined functions of covariates

Each model parameter of the univariate extreme value model has its own set of basis functions and coefficients.

Representation in Julia

In Julia, observations and covariates are stored in one DataFrame. The functional form of each parameter is specified using the @formula interface from StatsModels.jl. The formula defines how the transformed covariates enter the linear predictor for a given parameter. For example,

@formula(μ ~ 1 + x + log(z))

corresponds to a parameter function of the form $ \mu = \beta{0} + \beta{1}x + \beta_{2}\log(z). $

SpatialBayesExtremes.RegressionExtremeValueModel — Type

abstract type RegressionExtremeValueModel{D<:UnivariateDistribution} <: ContinuousExtremeValueModel end

An abstract type for regression extreme value models.

We consider the following regression models:

SpatialBayesExtremes.RegressionGEVModel — Type

RegressionGEVModel

A regression model for generalized extreme value distributed data where the location, scale, and shape parameters are modeled as functions of covariates. The model is suited for modeling the marginal behavior of block maxima with non-stationary characteristics.

fitSpatialgevML(data::Vector{<:Real}, covariates::DataFrame, locationFormula::FormulaTerm, scaleFormula::FormulaTerm, shapeFormula::FormulaTerm)                                        #maximum likelihood estimation
fitSpatialgevML(data::Vector{<:Real}, column::Symbol, locationFormula::FormulaTerm, scaleFormula::FormulaTerm, shapeFormula::FormulaTerm)                                               #maximum likelihood estimation from joint DataFrame for data and covariates, specifying column for data
fitSpatialgevML(data::Vector{<:Real}, covariates::DataFrame, locationFormula::FormulaTerm, scaleFormula::FormulaTerm, shapeFormula::FormulaTerm, initialvalue::NamedTuple)              #maximum likelihood estimation with initial values
fitSpatialgevML(data::Vector{<:Real}, column::Symbol, locationFormula::FormulaTerm, scaleFormula::FormulaTerm, shapeFormula::FormulaTerm, initialvalue::NamedTuple)                     #maximum likelihood estimation from joint DataFrame for data and covariates, specifying column for data, with initial values

SpatialBayesExtremes.RegressionGPDModel — Type

RegressionGPDModel

A regression generalized Pareto distribution (GPD) model for threshold exceedances with covariate-dependent scale and shape parameters. The model is designed to model the tail behavior of data exceeding a specified threshold while accounting for non-stationary characteristics through covariates.

fitSpatialgpdML(data::Vector{<:Real}, covariates::DataFrame, scaleFormula::FormulaTerm, shapeFormula::FormulaTerm)                                        #maximum likelihood estimation
fitSpatialgpdML(data::Vector{<:Real}, column::Symbol, scaleFormula::FormulaTerm, shapeFormula::FormulaTerm)                                               #maximum likelihood estimation from joint DataFrame for data and covariates, specifying column for data
fitSpatialgpdML(data::Vector{<:Real}, covariates::DataFrame, scaleFormula::FormulaTerm, shapeFormula::FormulaTerm, initialvalue::NamedTuple)              #maximum likelihood estimation with initial values
fitSpatialgpdML(data::Vector{<:Real}, column::Symbol, scaleFormula::FormulaTerm, shapeFormula::FormulaTerm, initialvalue::NamedTuple)                     #maximum likelihood estimation from joint DataFrame for data and covariates, specifying column for data, with initial values