Discrete Extreme Value Models

Any extreme value model that does not require any covariates and is thus restricted to extrapolation at the observation sites is considered a subtype of DiscreteExtremeValueModel.

SpatialBayesExtremes.DiscreteExtremeValueModel — Type

DiscreteExtremeValueModel <: ExtremeValueModel

An abstract type for discrete extreme value models.

Currently, the only available class of discrete extreme value models are univariate extreme value models.

Univariate Extreme Value Models

The abstract type UnivariateExtremeValueModel subsumes different parametric families of univariate distributions suitable to model univariate extremes.

SpatialBayesExtremes.UnivariateExtremeValueModel — Type

UnivariateExtremeValueModel{D<:UnivariateDistribution} <: DiscreteExtremeValueModel end

A UnivariateExtremeValueModel is any type of discrete extreme value model used to represent and analyze univariate extreme values in a dataset.

We consider the following classes of distributions:

SpatialBayesExtremes.UnivariateGumbelModel — Type

UnivariateGumbelModel

A univariate extreme value model for the Gumbel distribution suitable for modeling maxima of a dataset. The Gumbel distribution is characterized by its location parameter μ and scale parameter σ. We reparametrize the scale parameter using its logarithm to ensure positivity during estimation.

fitgumbelPWM(data::Vector{<:Real})                  #fit the univariate Gumbel model using PWM estimation
fitgumbelPWM(data::DataFrame, column::Symbol)       #fit the univariate Gumbel model using PWM estimation on a DataFrame column

External links

Gumbel distribution on Wikipedia

SpatialBayesExtremes.UnivariateGEVModel — Type

" UnivariateGEVModel

A univariate extreme value model for the Generalized Extreme Value (GEV) distribution suitable for modeling maxima of a dataset. The GEV distribution is characterized by its location parameter μ, scale parameter σ, and shape parameter ξ. We reparametrize the scale parameter using its logarithm to ensure positivity during estimation.

fitgevPWM(data::Vector{<:Real})                                            #fit the univariate GEV model using PWM estimation
fitgevPWM(data::DataFrame, column::Symbol)                                 #fit the univariate GEV model using PWM estimation on a DataFrame column
fitgevML(data::Vector{<:Real})                                             #fit the univariate GEV model using Maximum Likelihood estimation
fitgevML(data::DataFrame, column::Symbol)                                  #fit the univariate GEV model using Maximum Likelihood estimation on a DataFrame column
fitgevML(data::Vector{<:Real}, initialvalue::NamedTuple)                   #fit the univariate GEV model using Maximum Likelihood estimation with initial values (μ=μ0, σ=log(σ0), ξ=ξ0)
fitgevML(data::DataFrame, column::Symbol, initialvalue::NamedTuple)        #fit the univariate GEV model using Maximum Likelihood estimation on a DataFrame column with initial values (μ=μ0, σ=log(σ0), ξ=ξ0)

External links

Generalized extreme value distribution on Wikipedia

SpatialBayesExtremes.UnivariateGPDModel — Type

UnivariateGPDModel

A univariate extreme value model for the Generalized Pareto Distribution (GPD) suitable for modeling exceedances over a threshold. The GPD is characterized by its scale parameter σ and shape parameter ξ. We reparametrize the scale parameter using its logarithm to ensure positivity during estimation.

fitgpdPWM(data::Vector{<:Real})                                             #fit the univariate GPD model using PWM estimation
fitgpdPWM(data::DataFrame, column::Symbol)                                  #fit the univariate GPD model using PWM estimation on a DataFrame column
fitgpdHill(data::Vector{<:Real})                                            #fit the univariate GPD model using Hill estimation
fitgpdHill(data::DataFrame, column::Symbol)                                 #fit the univariate GPD model using Hill estimation on a DataFrame column
fitgpdMeplot(data::Vector{<:Real})                                          #fit the univariate GPD model using Mean Excess estimation
fitgpdMeplot(data::DataFrame, column::Symbol)                               #fit the univariate GPD model using Mean Excess estimation on a DataFrame column
fitgpdML(data::Vector{<:Real})                                              #fit the univariate GPD model using Maximum Likelihood estimation
fitgpdML(data::DataFrame, column::Symbol)                                   #fit the univariate GPD model using Maximum Likelihood estimation on a DataFrame column
fitgpdML(data::Vector{<:Real}, initialvalue::NamedTuple)                    #fit the univariate GPD model using Maximum Likelihood estimation with initial values (σ=log(σ0), ξ=ξ0)
fitgpdML(data::DataFrame, column::Symbol, initialvalue::NamedTuple)         #fit the univariate GPD model using Maximum Likelihood estimation on a DataFrame column with initial values (σ=log(σ0), ξ=ξ0)

External links

Generalized Pareto distribution on Wikipedia