Type Hierarchy

The extreme value models and fitted models are organized into a type hierarchy as follows:

Extreme Value Models

An extreme value model provides all the information and metadata necessary to apply an estimation method, such as the parametric family, the data and covariates.

ExtremeValueModel

The abstract type ExtremeValueModel subsumes any types of objects which specify a parametric family meant for estimation of extremes and all necessary sample information.

SpatialBayesExtremes.ExtremeValueModel — Type

abstract type ExtremeValueModel end

ExtremeValueModel is any type of statistical model used to represent and analyze extreme values in a dataset. Any ExtremeValueModel implements the following functions:

getDistribution

likelihood

loglikelihood

SpatialBayesExtremes.getDistribution — Method

getDistribution(model::ExtremeValueModel,θ::NamedTuple)

Get the probability distribution associated with the extreme value model given the parameters θ.

SpatialBayesExtremes.likelihood — Function

likelihood(model::ExtremeValueModel,θ::NamedTuple)

Compute the likelihood of the data linked to an extreme value model given the parameters θ.

SpatialBayesExtremes.loglikelihood — Function

loglikelihood(model::ExtremeValueModel,θ::NamedTuple)

Compute the loglikelihood of the data linked to an extreme value model given the parameters θ.

Currently, we distinguish between two subtypes of ExtremeValueModel, depending on whether the model includes covariates and allows for extrapolation beyond the sample covariates (continuous) or not (discrete).

abstract type DiscreteExtremeValueModel <: ExtremeValueModel end
abstract type ContinuousExtremeValueModel <: ExtremeValueModel end

Fitted Estimators

The fitted estimator yields the connection between statistical model and the estimation method generating the parameter estimate.

FittedEstimator

Any type of object linking a ExtremeValueModel to a estimation method and the parameter estimates is a subtype of the abstract type FittedEstimator.

abstract type FittedEstimator{M<:ExtremeValueModel} end

It has one parameter that specifies the statistical model it is the parameter estimates correspond to.

SpatialBayesExtremes.FittedEstimator — Type

FittedEstimator{M<:ExtremeValueModel}

An abstract type for fitted estimators of extreme value models.

SpatialBayesExtremes.getParameterEstimate — Function

getParameterEstimate(fe::FittedEstimator)

Returns the parameter estimates.

SpatialBayesExtremes.getDistributionEstimate — Function

getDistributionEstimate(fe::FittedEstimator)

Returns the probability distribution associated with the fitted estimator.

Each concrete type corresponds to one estimation method provided by this package. Currently, the following estimation methods exist:

Probability-weighted moment estimation for Gumbel distribution, genralized extreme value distribution and generalized Pareto distribution

Mean excess based Hill estimation and mean excess estimator for generalized Pareto distribution

Maximum likelihood estimation